Remove references to App Center
App Center has been discontinued

We've replaced it with Firebase App Distribution for iOS builds, but FAD does not seem to support macOS binaries. So instead we are keeping the lane to build the app with DeveloperID signing, but will let devs decide how to distribute it for the time being.

Using `build_and_upload_app_store` to upload the builds to TestFlight and use TestFlight for beta testing is now the preferred way to distribute betas.

If we end up really needing to distribute separate _internal_ betas of SNMac signed with DeveloperID on a frequent-ish basis in the future, we could consider still using this `build_beta` lane and distributing them through AppsCDN; but at that point it's not a need we have anymore.